pyproject.toml
[tool.poetry]
name = "twitterbot"
version = "1.0.0"
description = ""
authors = ["some guy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.14.0"
tracery = "^0.1.1"
flask = "^2.3.2"
python-dotenv = "^1.0.0"
flask-login = "^0.6.2"
openai = "^0.27.6"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
